Wiredash is a great feedback tool
A new flutter project has been created to test out a feedback method for comfy Space.

A user will give feedback , and I need that experience to be good.

Popular options are:
1. Email -> insecure
2. Firebase -> a pain to set up and still much work needed done

And I found wiredash, specifically made for this.
Kid you not, it's only like 5 lines of code.

Great tool