New experiment with SQFLITE + Stream builder
The current state of comfy space is quite laggy in the in-between.
I want that to be nice & responsive. So we're moving from future buildter to stream builder to not have to rerender the whole thing.

With this, also we might want to change from sqflite? moor and hive sounds cool